requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cohere==4.32
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dacite==1.8.1
dataclasses-json==0.6.1
distro==1.8.0
fastapi==0.104.0
fastavro==1.8.2
ffmpy==0.3.1
filelock==3.13.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.0.1
h11==0.14.0
htmlmin==0.1.12
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.0.325
langsmith==0.0.53
llvmlite==0.41.1
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.3
multidict==6.0.4
multimethod==1.10
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
networkx==3.2.1
numba==0.58.1
numpy==1.25.2
openai==1.3.7
orjson==3.9.10
packaging==23.2
pandas==2.0.3
patsy==0.5.3
phik==0.12.3
Pillow==10.1.0
plotly==5.18.0
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rpds-py==0.10.6
scipy==1.11.3
seaborn==0.12.2
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
statsmodels==0.14.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
tiktoken==0.5.1
toolz==0.12.0
tqdm==4.66.1
typeguard==4.1.5
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
visions==0.7.5
websockets==11.0.3
wordcloud==1.9.2
yarl==1.9.2
yaspin==3.0.1
ydata-profiling==4.6.1
zipp==3.17.0




